reuse OkHttpClient instance
